chore: extract shared search paths code from modman
